Tests fixed

+++ tests/test_message_frame.cpp
@@ -8,31 +8,31 @@ using msgframe::MessageFrame;
 using msgframe::ParameterValue;
 
 // --------------------------------------------------------------------
-// Group: Serialization — Перевірка бінарного пакінгу та heap-reuse
+// Group: Serialization — Checking binary packing and heap-reuse
 // --------------------------------------------------------------------
 
 TEST(Serialization, EndToEndPackUnpackWithHeavyAttachments) {
     MessageFrame source_frame;
 
-    // Наповнюємо метаданими
-    source_frame.add("system", "status", ParameterValue(std::string("operational")));
-    source_frame.add("sensor", "reading", ParameterValue(3.14159));
-
-    // Додаємо важкий сирий бінарний атачмент (наприклад, масив сирих сигналів з SDR)
-    std::vector<uint8_t> mock_signal(2048, 0xAB);
-    source_frame.add_attachment("raw_sdr_channel_A", std::move(mock_signal));
-
-    // Серіалізація в повторно використовуваний буфер
-    std::vector<uint8_t> buffer;
-    source_frame.serialize(buffer);
-    CHECK(buffer.size() > 0);
-
-    // Десеріалізація в абсолютно новий фрейм
-    MessageFrame target_frame;
-    bool unpack_ok = target_frame.deserialize(buffer.data(), buffer.size());
-    CHECK(unpack_ok);
-
-    // Перевірка консистентності метаданих після мережевого розпакування
+	// Fill with metadata
+	source_frame.add("system", "status", ParameterValue(std::string("operational")));
+	source_frame.add("sensor", "reading", ParameterValue(3.14159));
+	
+	// Add a heavy raw binary attachment (e.g., an array of raw signals from an SDR)
+	std::vector<uint8_t> mock_signal(2048, 0xAB);
+	source_frame.add_attachment("raw_sdr_channel_A", std::move(mock_signal));
+	
+	// Serialize to a reusable buffer
+	std::vector<uint8_t> buffer;
+	source_frame.serialize(buffer);
+	CHECK(buffer.size() > 0);
+	
+	// Deserialize to a completely new frame
+	MessageFrame target_frame;
+	bool unpack_ok = target_frame.deserialize(buffer.data(), buffer.size());
+	CHECK(unpack_ok);
+	
+	// Check metadata consistency after network unpacking
     const auto* status = target_frame.find("system", "status");
     CHECK_NOT_NULL(status);
     if (status) {
@@ -41,14 +41,14 @@ TEST(Serialization, EndToEndPackUnpackWithHeavyAttachments) {
         if (s) CHECK_EQ(*s, std::string("operational"));
     }
 
-    // Перевірка збереження бінарних атачментів через індекси вектора
+    // Checking the preservation of binary attachments via vector indices
     const auto& target_attachments = target_frame.get_attachments();
     CHECK_EQ(target_attachments.size(), static_cast<size_t>(1));
     if (!target_attachments.empty()) {
         CHECK_EQ(target_attachments[0].name, std::string("raw_sdr_channel_A"));
         CHECK_EQ(target_attachments[0].raw_data.size(), static_cast<size_t>(2048));
 
-        // Поелементна перевірка байтів буфера
+        // Element-by-element inspection of buffer bytes
         CHECK_EQ(target_attachments[0].raw_data[0], static_cast<uint8_t>(0xAB));
         CHECK_EQ(target_attachments[0].raw_data[2047], static_cast<uint8_t>(0xAB));
     }
@@ -56,33 +56,35 @@ TEST(Serialization, EndToEndPackUnpackWithHeavyAttachments) {
 
 TEST(Serialization, CapacityRetentionOnConsecutiveSerializations) {
     MessageFrame msg;
-    msg.add("dev", "p", ParameterValue(42));
+    msg.add("dev", "p", ParameterValue(static_cast<int64_t>(42)));
 
     std::vector<uint8_t> shared_network_buffer;
-    shared_network_buffer.reserve(4096); // Пре-аллокація пулу
+    shared_network_buffer.reserve(4096); // Pool pre-allocation
 
     msg.serialize(shared_network_buffer);
     size_t cap_after_first = shared_network_buffer.capacity();
 
-    // Перевіряємо, що VectorBuffer не зламав capacity нашого мережевого пулу
+    // Check that VectorBuffer has not broken the capacity of our network pool
     CHECK(cap_after_first >= static_cast<size_t>(4096));
 
-    // Друга серіалізація в цей самий буфер не повинна провокувати нові алокації
+    // A second serialization to the same buffer should not provoke new allocations.
     msg.serialize(shared_network_buffer);
     CHECK_EQ(shared_network_buffer.capacity(), cap_after_first);
 }
 
 TEST(Serialization, MalformedBinaryBufferReturnsFalseGracefully) {
     MessageFrame msg;
-    // Підсовуємо десеріалізатору випадкове бінарне сміття замість MessagePack
+    // Feed the deserializer random binary garbage instead of MessagePack
     std::vector<uint8_t> garbage(50, 0xFE);
 
     bool result = msg.deserialize(garbage.data(), garbage.size());
 
-    // Програма не повинна падати по Segmentation Fault; unpack має повернути false
+    // The program should not crash on a Segmentation Fault; unpack should return false
     CHECK_FALSE(result);
 }
 
+/*
 int main() {
     return msgframe_test::run_all();
 }
+*/
